refactor schemePaymentsByYearModel
