+++ src/routes/models/scheme-payments-by-year/model.js
@@ -1,0 +1,8 @@
+import { fetchSchemePaymentsByYear } from './fetch.js'
+import { transformSchemePaymentsData } from './transform.js'
+
+export async function schemePaymentsByYearModel () {
+  const rawData = await fetchSchemePaymentsByYear()
+
+  return transformSchemePaymentsData(rawData)
+}
