+++ src/routes/scheme-payments-by-year.js
@@ -1,4 +1,4 @@
-import { schemePaymentsByYearModel } from './models/scheme-payments-by-year.js'
+import { schemePaymentsByYearModel } from './models/scheme-payments-by-year/model.js'
 
 export const schemePaymentsByYear = {
   method: 'GET',
